git-svn-id: http://192.168.20.11/sequoiadb/trunk@11772 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/test/clientTest.cpp
@@ -343,7 +343,7 @@ INT32 msgReceive ()
          goto error ;
       }
       header = (MsgHeader*)pInBuffer ;
-      if ( header->requestID != requestID )
+      if ( header->requestID != (UINT64)requestID )
       {
          printf ( "-Warning: Reply request id is not the same, expect:%lld, "
                   "recieved:%lld\n", requestID, header->requestID ) ;
